pyproject.toml
[tool.poetry]
name = "honkers"
version = "0.1.0"
description = ""
authors = ["Nights"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
replit = "^2.0.0"
flask = "^1.1.2"
pymongo = "^3.11.3"
dnspython = "^2.1.0"
dadjokes = "^1.3.2"
discord-pretty-help = "^1.2.2"
pytz = "^2021.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
